analisis grafik final

+++ app/src/main/java/com/example/appmoneego/model/KategoriSummary.kt
@@ -3,5 +3,5 @@ package com.example.appmoneego.model
 data class KategoriSummary(
     val nama: String,
     val jumlah: Double,
-    val persentase: Double
+    val persentase: Float
 )
